cogl: Mark public symbols for export
This way we can reduce the maintenance burden for updating the
cogl.symbols files for the Visual Studio builds, as we can now export
symbols using __declspec(dllexport).  The cogl.symbols file is still
kept to remind us of the usage of private Cogl symbols in Cogl-Path
and Cogl-Pango that we want to replace with public with public symbols.