cogl-crate.c: Fix running on 64-bit Windows
CoglPollFD uses an int type for its fd member but the GPollFD that
g_poll() expects uses a gint64 type (a.k.a int64_t, __in64 on Windows)
for its fd member on 64-bit Windows, which will cause loads of
"WaitForMultipleObjectsEx failed:..." to be spewed from GLib on 64-bit
Windows.

This updates the example to use the correct types for the g_poll() call,
which will avoid this kind of situation.